Basically only imortant from #408
Leave blackSqare and step counters alone, but make it be able to handle textures where width and height are not the same so no more invalid color error.
For example, with original code, materials.newDebugTexture(8, 16, 40) has odd color because when it gets to pixel (1, 9), color value is (15, 9, -2) in RGB. After color.RGBToInteger inside color.HSBToInteger, integer value for color would be -2. The reason it's not apparent is likely from optimization of colors from palette, turning -2 into 65535 or (0, 255, 255), instead of an invalid color.
This has the effect of screwing up the colors for the image
If steps are swapped as this proposes, this issue is fixed and the colors wouldn't go crazy.